requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
matplotlib>=3.4.0
numpy>=1.21.0
Pillow>=9.0.0
h5py>=3.7.0
scipy>=1.7.0
opencv-python>=4.5.0
glob2>=0.7
scikit-image>=0.18.0
